requirements.txt
requests==2.11.1
python-whois==0.6.7
